+++ sdk/CHANGELOG.md
@@ -38,6 +38,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `SecretManage::sign_evm`;
 - `Account::addresses_balance` method accepting addresses to get balance for;
 - `Topic::is_valid`;
+- `Wallet::get_secret_manager` method;
 
 ### Changed